@@ -13,20 +13,27 @@ classifiers =
     Intended Audience :: Developers
     License :: OSI Approved :: MIT License
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
 
 [options]
 include_package_data = True
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.9
 install_requires =
     dragodis>=0.8.0
-    click
+    cmd2>=2
+    pyreadline3; sys_platform == "win32"
+    tabulate
     hexdump
-    pyhidra>=0.3.0
+    pyhidra>=0.6.0
     yara-python
 
 [options.entry_points]
+console_scripts =
+    rugosa = rugosa.shell:main
 pyhidra.setup =
     rugosa = rugosa.ghidra_plugin.install:setup
 pyhidra.pre_launch =
